完善 请求任务和日志

+++ services/daemon.py
@@ -9,27 +9,30 @@ import time
 import traceback
 
 import psutil
-
 from config import BASE_DIR
 from logger import BaseLog
-from services.screen import start_screen
+from services.screen import start_screen, start_cmd
 
 daemon_log = BaseLog('services', 'daemon')
 
-proc_info_list = []
+_proc_info_list = []
+_start_cmd_dict = {
+    'screen': start_cmd,
+    'crawler': '{} main.py'.format(sys.executable, os.path.join(BASE_DIR, 'main.py')),
+    'daemon': '{} {}'.format(sys.executable, os.path.join(BASE_DIR, 'daemon.py')),
+}
 
 
 def daemon():
-    time.sleep(1)
     c_pid = crawler_thread_start()
     s_pid = screen_thread_start()
-    proc_info_list.extend([psutil.Process(c_pid).as_dict(attrs=['exe', 'cmdline', 'pid']),
-                           psutil.Process(s_pid).as_dict(attrs=['exe', 'cmdline', 'pid'])])
+    _proc_info_list.extend([psutil.Process(c_pid).as_dict(attrs=['exe', 'cmdline', 'pid']),
+                                psutil.Process(s_pid).as_dict(attrs=['exe', 'cmdline', 'pid'])])
     while True:
         time.sleep(5)
         try:
-            for i in range(len(proc_info_list)):
-                proc_info = proc_info_list[i]
+            for i in range(len(_proc_info_list)):
+                proc_info = _proc_info_list[i]
                 try:
                     _proc = psutil.Process(proc_info.get('pid'))
                     cmd = _proc.cmdline()
@@ -47,14 +50,15 @@ def daemon():
                 if pid == 0:
                     break
                 temp = psutil.Process(pid).as_dict(attrs=['exe', 'cmdline', 'pid'])
-                proc_info_list[i] = temp
+                _proc_info_list[i] = temp
         except Exception:
             daemon_log.error(traceback.format_exc())
 
 
 def start_daemon():
     try:
-        cmd = '{} {}'.format(sys.executable, os.path.join(BASE_DIR, 'daemon.py'))
+        cmd = _start_cmd_dict.get('daemon')
+        print(cmd)
         s = subprocess.Popen(cmd, creationflags=subprocess.CREATE_NEW_CONSOLE, universal_newlines=True,
                              encoding='utf-8', shell=True)
         time.sleep(5)
@@ -66,7 +70,7 @@ def start_daemon():
                 break
         p.kill()
         if crawler_pid != 0:
-            print('daemon Process success start * * * pid-[{}]'.format(crawler_pid))
+            print('daemon Process start success  * * * pid-[{}]'.format(crawler_pid))
         else:
             print('daemon Process start error * * *')
     except Exception as e:
@@ -74,11 +78,8 @@ def start_daemon():
 
 
 def crawler_thread_start():
-    crawler_path = os.path.join(BASE_DIR, 'main.py')
-    cmd = '{} main.py'.format(sys.executable, crawler_path)
-    s = subprocess.Popen(cmd, creationflags=subprocess.CREATE_NEW_CONSOLE,
-                         universal_newlines=True, encoding='utf-8',
-                         shell=True)
+    cmd = _start_cmd_dict.get('crawler')
+    s = subprocess.Popen(cmd, creationflags=subprocess.CREATE_NEW_CONSOLE, shell=True)
     time.sleep(3)
     p = psutil.Process(s.pid)
     crawler_pid = 0
@@ -103,25 +104,40 @@ def screen_thread_start():
     return screen_pid
 
 
-def crawler_stop():
-    p = psutil.Process(proc_info_list[0].get('pid'))
-    p.kill()
+def stop(proc_name: str = ''):
+    _cmdline = []
+    if _start_cmd_dict.get(proc_name, ''):
+        stop = True
+        _cmdline.append(_start_cmd_dict.get(proc_name, ''))
+    else:
+        stop = False
+        _cmdline.extend(list(_start_cmd_dict.values()))
+    for proc in psutil.process_iter():
+        _proc = proc.as_dict(attrs=['exe', 'cmdline', 'pid'])
+        cmdlind = _proc.get('cmdline')
+        for i in _cmdline:
+            if cmdlind and i == ' '.join(cmdlind):
+                pid = _proc.get('pid')
+                try:
+                    p = psutil.Process(pid)
+                    p.kill()
+                    # _proc_info_list.remove(_proc)
+                    print('{} Process kill success'.format(cmdlind))
+                except psutil.NoSuchProcess:
+                    print("{} 进程已停止运行")
+                if stop:
+                    return
 
 
-def screen_stop():
-    p = psutil.Process(proc_info_list[1].get('pid'))
-    p.kill()
-    pass
-
-
-def deamon_stop(pid):
-    p = psutil.Process(pid)
-    p.kill()
-
+def restart():
+    stop()
+    print('stop success * * *')
+    start_daemon()
+    print('restart success * * *')
 
 def statu():
-    p1 = psutil.Process(proc_info_list[0].get('pid'))
-    p2 = psutil.Process(proc_info_list[1].get('pid'))
+    p1 = psutil.Process(_proc_info_list[0].get('pid'))
+    p2 = psutil.Process(_proc_info_list[1].get('pid'))
     print("crawler is running:{} ****".format(p1.is_running()))
     print("screen is running:{} ****".format(p2.is_running()))
